appendix 2. combinate keyword.
seperated excel files are re-loaded one more time. then combinated 1 or 2 keywords. 
more than 100 frequencies, then that combiation can be an episode. or? deleted. that is 2_2_100.R
But, based on first method, a lot of responses are deleted, and the thing critical but not frequent things(e.g. suicide)are excluded. So, in order to check episodes in hand, 2_2_all codes are used.